resolves gh-4 without state; optimize key listener to use event.key; rename component to VirtualKeyboard
